raster working with gaps

+++ 2.0/libs/extra/RenderUtils.h
@@ -6,6 +6,6 @@
 
 class RenderUtils {
 public:
-    static std::vector<CanvasPoint> interpolateCanvasPoints(CanvasPoint from, CanvasPoint to, size_t numberOfValues);
-    static std::vector<TexturePoint> interpolateTexturePoints(TexturePoint from, TexturePoint to, size_t numberOfValues);
+    static std::vector<CanvasPoint> interpolateCanvasPoints(CanvasPoint from, CanvasPoint to, float numberOfValues);
+    static std::vector<TexturePoint> interpolateTexturePoints(TexturePoint from, TexturePoint to, float numberOfValues);
 };
